rename stpp test resources
