+++ test/manifests/stpp-no-init.js
