+++ test/manifests/stpp-no-init.mpd
